Move creation to io
