+++ ext/io/Cargo.toml
@@ -25,8 +25,10 @@ tokio.workspace = true
 
 [target.'cfg(not(windows))'.dependencies]
 os_pipe.workspace = true
+libc.workspace = true
 
 [target.'cfg(windows)'.dependencies]
 winapi = { workspace = true, features = ["winbase", "processenv", "errhandlingapi"] }
 rand.workspace = true
 parking_lot.workspace = true
+windows-sys.workspace = true
